+++ src/NewPicker/PickerPanel/DatePanel/index.tsx
@@ -8,7 +8,7 @@ import {
   WEEK_DAY_COUNT,
 } from '../../../utils/dateUtil';
 import type { PanelMode, SharedPanelProps } from '../../../interface';
-import { PanelContext, useInfo } from '../context';
+import { PanelContext, useInfo } from '../../../PickerPanel/context';
 import PanelBody from '../../../PickerPanel/PanelBody';
 import PanelHeader from '../../../PickerPanel/PanelHeader';
