+++ src/NewPicker/PickerPanel/MonthPanel/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { formatValue, isSameMonth } from '../../../utils/dateUtil';
 import type { SharedPanelProps } from '../../../interface';
-import { PanelContext, useInfo } from '../context';
+import { PanelContext, useInfo } from '../../../PickerPanel/context';
 import PanelBody from '../../../PickerPanel/PanelBody';
 import PanelHeader from '../../../PickerPanel/PanelHeader';
