+++ src/NewPicker/PickerPanel/YearPanel/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import type { SharedPanelProps } from '../../../interface';
 import { formatValue, isSameYear } from '../../../utils/dateUtil';
-import { PanelContext, useInfo } from '../context';
+import { PanelContext, useInfo } from '../../../PickerPanel/context';
 import PanelBody from '../../../PickerPanel/PanelBody';
 import PanelHeader from '../../../PickerPanel/PanelHeader';
